+++ dpcpp/test/matrix/CMakeLists.txt
@@ -1,4 +1,3 @@
-ginkgo_create_test(coo_kernels)
 ginkgo_create_test(csr_kernels)
 ginkgo_create_test(dense_kernels)
 ginkgo_create_test(diagonal_kernels)
