+++ hip/test/matrix/CMakeLists.txt
@@ -1,4 +1,3 @@
-ginkgo_create_hip_test(coo_kernels)
 ginkgo_create_hip_test(csr_kernels)
 ginkgo_create_hip_test(dense_kernels)
 ginkgo_create_hip_test(diagonal_kernels)
