+++ test/matrix/coo_kernels.cpp
@@ -50,79 +50,50 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "core/matrix/coo_kernels.hpp"
 #include "core/test/utils.hpp"
 #include "core/test/utils/unsort_matrix.hpp"
+#include "test/utils/executor.hpp"
 
 
-namespace {
-
-
-class Coo : public ::testing::Test {
+class Coo : public CommonTestFixture {
 protected:
-    using Mtx = gko::matrix::Coo<>;
-    using Vec = gko::matrix::Dense<>;
-    using ComplexVec = gko::matrix::Dense<std::complex<double>>;
+#if GINKGO_COMMON_SINGLE_MODE
+    using vtype = float;
+#else
+    using vtype = double;
+#endif  // GINKGO_COMMON_SINGLE_MODE
+    using Mtx = gko::matrix::Coo<vtype>;
+    using Vec = gko::matrix::Dense<vtype>;
+    using ComplexVec = gko::matrix::Dense<std::complex<vtype>>;
 
-    Coo() : mtx_size(532, 231), rand_engine(42) {}
-
-    void SetUp()
-    {
-        ref = gko::ReferenceExecutor::create();
-        omp = gko::OmpExecutor::create();
-    }
-
-    void TearDown()
-    {
-        if (omp != nullptr) {
-            ASSERT_NO_THROW(omp->synchronize());
-        }
-    }
+    Coo() : rand_engine(42) {}
 
     template <typename MtxType = Vec>
-    std::unique_ptr<MtxType> gen_mtx(int num_rows, int num_cols,
-                                     int min_nnz_row)
+    std::unique_ptr<MtxType> gen_mtx(int num_rows, int num_cols)
     {
         return gko::test::generate_random_matrix<MtxType>(
-            num_rows, num_cols,
-            std::uniform_int_distribution<>(min_nnz_row, num_cols),
-            std::normal_distribution<>(-1.0, 1.0), rand_engine, ref);
+            num_rows, num_cols, std::uniform_int_distribution<>(1, num_cols),
+            std::normal_distribution<vtype>(-1.0, 1.0), rand_engine, ref);
     }
 
     void set_up_apply_data(int num_vectors = 1)
     {
-        mtx = gen_mtx<Mtx>(mtx_size[0], mtx_size[1], 1);
-        expected = gen_mtx(mtx_size[0], num_vectors, 1);
-        y = gen_mtx(mtx_size[1], num_vectors, 1);
+        mtx = gen_mtx<Mtx>(532, 231);
+        expected = gen_mtx(532, num_vectors);
+        y = gen_mtx(231, num_vectors);
         alpha = gko::initialize<Vec>({2.0}, ref);
         beta = gko::initialize<Vec>({-1.0}, ref);
-        dmtx = gko::clone(omp, mtx);
-        dresult = gko::clone(omp, expected);
-        dy = gko::clone(omp, y);
-        dalpha = gko::clone(omp, alpha);
-        dbeta = gko::clone(omp, beta);
+        dmtx = gko::clone(exec, mtx);
+        dresult = gko::clone(exec, expected);
+        dy = gko::clone(exec, y);
+        dalpha = gko::clone(exec, alpha);
+        dbeta = gko::clone(exec, beta);
     }
 
-    struct matrix_pair {
-        std::unique_ptr<Mtx> ref;
-        std::unique_ptr<Mtx> omp;
-    };
-
-    matrix_pair gen_unsorted_mtx()
+    void unsort_mtx()
     {
-        constexpr int min_nnz_per_row{2};
-        auto local_mtx_ref = Mtx::create(ref);
-        auto generated = gen_mtx(mtx_size[0], mtx_size[1], min_nnz_per_row);
-        local_mtx_ref->copy_from(generated.get());
-        gko::test::unsort_matrix(local_mtx_ref.get(), rand_engine);
-
-        auto local_mtx_omp = gko::clone(omp, local_mtx_ref);
-
-        return {std::move(local_mtx_ref), std::move(local_mtx_omp)};
+        gko::test::unsort_matrix(mtx.get(), rand_engine);
+        dmtx->copy_from(mtx.get());
     }
 
-
-    std::shared_ptr<gko::ReferenceExecutor> ref;
-    std::shared_ptr<const gko::OmpExecutor> omp;
-
-    const gko::dim<2> mtx_size;
     std::default_random_engine rand_engine;
 
     std::unique_ptr<Mtx> mtx;
@@ -146,7 +117,7 @@ TEST_F(Coo, SimpleApplyIsEquivalentToRef)
     mtx->apply(y.get(), expected.get());
     dmtx->apply(dy.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
@@ -154,15 +125,29 @@ TEST_F(Coo, SimpleApplyDoesntOverwritePadding)
 {
     set_up_apply_data();
     auto dresult_padded =
-        Vec::create(omp, dresult->get_size(), dresult->get_stride() + 1);
+        Vec::create(exec, dresult->get_size(), dresult->get_stride() + 1);
     dresult_padded->copy_from(dresult.get());
-    dresult_padded->get_values()[1] = 1234.0;
+    vtype padding_val{1234.0};
+    exec->copy_from(exec->get_master().get(), 1, &padding_val,
+                    dresult_padded->get_values() + 1);
 
     mtx->apply(y.get(), expected.get());
     dmtx->apply(dy.get(), dresult_padded.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult_padded, expected, 1e-14);
-    ASSERT_EQ(dresult_padded->get_values()[1], 1234.0);
+    GKO_ASSERT_MTX_NEAR(dresult_padded, expected, r<vtype>::value);
+    ASSERT_EQ(exec->copy_val_to_host(dresult_padded->get_values() + 1), 1234.0);
+}
+
+
+TEST_F(Coo, SimpleApplyIsEquivalentToRefUnsorted)
+{
+    set_up_apply_data();
+    unsort_mtx();
+
+    mtx->apply(y.get(), expected.get());
+    dmtx->apply(dy.get(), dresult.get());
+
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
@@ -173,7 +158,7 @@ TEST_F(Coo, AdvancedApplyIsEquivalentToRef)
     mtx->apply(alpha.get(), y.get(), beta.get(), expected.get());
     dmtx->apply(dalpha.get(), dy.get(), dbeta.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
@@ -181,15 +166,17 @@ TEST_F(Coo, AdvancedApplyDoesntOverwritePadding)
 {
     set_up_apply_data();
     auto dresult_padded =
-        Vec::create(omp, dresult->get_size(), dresult->get_stride() + 1);
+        Vec::create(exec, dresult->get_size(), dresult->get_stride() + 1);
     dresult_padded->copy_from(dresult.get());
-    dresult_padded->get_values()[1] = 1234.0;
+    vtype padding_val{1234.0};
+    exec->copy_from(exec->get_master().get(), 1, &padding_val,
+                    dresult_padded->get_values() + 1);
 
     mtx->apply(alpha.get(), y.get(), beta.get(), expected.get());
     dmtx->apply(dalpha.get(), dy.get(), dbeta.get(), dresult_padded.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult_padded, expected, 1e-14);
-    ASSERT_EQ(dresult_padded->get_values()[1], 1234.0);
+    GKO_ASSERT_MTX_NEAR(dresult_padded, expected, r<vtype>::value);
+    ASSERT_EQ(exec->copy_val_to_host(dresult_padded->get_values() + 1), 1234.0);
 }
 
 
@@ -200,7 +187,7 @@ TEST_F(Coo, SimpleApplyAddIsEquivalentToRef)
     mtx->apply2(y.get(), expected.get());
     dmtx->apply2(dy.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
@@ -211,7 +198,7 @@ TEST_F(Coo, AdvancedApplyAddIsEquivalentToRef)
     mtx->apply2(alpha.get(), y.get(), expected.get());
     dmtx->apply2(dalpha.get(), dy.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
@@ -222,7 +209,7 @@ TEST_F(Coo, SimpleApplyToDenseMatrixIsEquivalentToRef)
     mtx->apply(y.get(), expected.get());
     dmtx->apply(dy.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
@@ -233,7 +220,7 @@ TEST_F(Coo, AdvancedApplyToDenseMatrixIsEquivalentToRef)
     mtx->apply(alpha.get(), y.get(), beta.get(), expected.get());
     dmtx->apply(dalpha.get(), dy.get(), dbeta.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
@@ -244,19 +231,30 @@ TEST_F(Coo, SimpleApplyAddToDenseMatrixIsEquivalentToRef)
     mtx->apply2(y.get(), expected.get());
     dmtx->apply2(dy.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
 TEST_F(Coo, SimpleApplyAddToDenseMatrixIsEquivalentToRefUnsorted)
 {
     set_up_apply_data(6);
-    auto pair = gen_unsorted_mtx();
+    unsort_mtx();
 
-    pair.ref->apply2(y.get(), expected.get());
-    pair.omp->apply2(dy.get(), dresult.get());
+    mtx->apply2(y.get(), expected.get());
+    dmtx->apply2(dy.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
+}
+
+
+TEST_F(Coo, SimpleApplyAddToLargeDenseMatrixIsEquivalentToRef)
+{
+    set_up_apply_data(33);
+
+    mtx->apply2(y.get(), expected.get());
+    dmtx->apply2(dy.get(), dresult.get());
+
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
@@ -267,62 +265,88 @@ TEST_F(Coo, AdvancedApplyAddToDenseMatrixIsEquivalentToRef)
     mtx->apply2(alpha.get(), y.get(), expected.get());
     dmtx->apply2(dalpha.get(), dy.get(), dresult.get());
 
-    GKO_ASSERT_MTX_NEAR(dresult, expected, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
+}
+
+
+TEST_F(Coo, AdvancedApplyAddToLargeDenseMatrixIsEquivalentToRef)
+{
+    set_up_apply_data(33);
+
+    mtx->apply2(y.get(), expected.get());
+    dmtx->apply2(dy.get(), dresult.get());
+
+    GKO_ASSERT_MTX_NEAR(dresult, expected, r<vtype>::value);
 }
 
 
 TEST_F(Coo, ApplyToComplexIsEquivalentToRef)
 {
     set_up_apply_data();
-    auto complex_b = gen_mtx<ComplexVec>(231, 3, 1);
-    auto dcomplex_b = gko::clone(omp, complex_b);
-    auto complex_x = gen_mtx<ComplexVec>(532, 3, 1);
-    auto dcomplex_x = gko::clone(omp, complex_x);
+    auto complex_b = gen_mtx<ComplexVec>(231, 3);
+    auto dcomplex_b = gko::clone(exec, complex_b);
+    auto complex_x = gen_mtx<ComplexVec>(532, 3);
+    auto dcomplex_x = gko::clone(exec, complex_x);
 
     mtx->apply(complex_b.get(), complex_x.get());
     dmtx->apply(dcomplex_b.get(), dcomplex_x.get());
 
-    GKO_ASSERT_MTX_NEAR(dcomplex_x, complex_x, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dcomplex_x, complex_x, r<vtype>::value);
 }
 
 
 TEST_F(Coo, AdvancedApplyToComplexIsEquivalentToRef)
 {
     set_up_apply_data();
-    auto complex_b = gen_mtx<ComplexVec>(231, 3, 1);
-    auto dcomplex_b = gko::clone(omp, complex_b);
-    auto complex_x = gen_mtx<ComplexVec>(532, 3, 1);
-    auto dcomplex_x = gko::clone(omp, complex_x);
+    auto complex_b = gen_mtx<ComplexVec>(231, 3);
+    auto dcomplex_b = gko::clone(exec, complex_b);
+    auto complex_x = gen_mtx<ComplexVec>(532, 3);
+    auto dcomplex_x = gko::clone(exec, complex_x);
 
     mtx->apply(alpha.get(), complex_b.get(), beta.get(), complex_x.get());
     dmtx->apply(dalpha.get(), dcomplex_b.get(), dbeta.get(), dcomplex_x.get());
 
-    GKO_ASSERT_MTX_NEAR(dcomplex_x, complex_x, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dcomplex_x, complex_x, r<vtype>::value);
 }
 
 
 TEST_F(Coo, ApplyAddToComplexIsEquivalentToRef)
 {
     set_up_apply_data();
-    auto complex_b = gen_mtx<ComplexVec>(231, 3, 1);
-    auto dcomplex_b = gko::clone(omp, complex_b);
-    auto complex_x = gen_mtx<ComplexVec>(532, 3, 1);
-    auto dcomplex_x = gko::clone(omp, complex_x);
+    auto complex_b = gen_mtx<ComplexVec>(231, 3);
+    auto dcomplex_b = gko::clone(exec, complex_b);
+    auto complex_x = gen_mtx<ComplexVec>(532, 3);
+    auto dcomplex_x = gko::clone(exec, complex_x);
 
     mtx->apply2(alpha.get(), complex_b.get(), complex_x.get());
     dmtx->apply2(dalpha.get(), dcomplex_b.get(), dcomplex_x.get());
 
-    GKO_ASSERT_MTX_NEAR(dcomplex_x, complex_x, 1e-14);
+    GKO_ASSERT_MTX_NEAR(dcomplex_x, complex_x, r<vtype>::value);
+}
+
+
+TEST_F(Coo, ConvertToDenseIsEquivalentToRef)
+{
+    set_up_apply_data();
+    auto dense_mtx = gko::matrix::Dense<vtype>::create(ref);
+    auto ddense_mtx = gko::matrix::Dense<vtype>::create(exec);
+
+    mtx->convert_to(dense_mtx.get());
+    dmtx->convert_to(ddense_mtx.get());
+
+    GKO_ASSERT_MTX_NEAR(dense_mtx.get(), ddense_mtx.get(), 0);
 }
 
 
 TEST_F(Coo, ConvertToCsrIsEquivalentToRef)
 {
     set_up_apply_data();
-    auto csr_mtx = gko::matrix::Csr<>::create(ref);
-    auto dcsr_mtx = gko::matrix::Csr<>::create(omp);
+    auto dense_mtx = gko::matrix::Dense<vtype>::create(ref);
+    auto csr_mtx = gko::matrix::Csr<vtype>::create(ref);
+    auto dcsr_mtx = gko::matrix::Csr<vtype>::create(exec);
 
-    mtx->convert_to(csr_mtx.get());
+    mtx->convert_to(dense_mtx.get());
+    dense_mtx->convert_to(csr_mtx.get());
     dmtx->convert_to(dcsr_mtx.get());
 
     GKO_ASSERT_MTX_NEAR(csr_mtx.get(), dcsr_mtx.get(), 0);
@@ -347,7 +371,7 @@ TEST_F(Coo, InplaceAbsoluteMatrixIsEquivalentToRef)
     mtx->compute_absolute_inplace();
     dmtx->compute_absolute_inplace();
 
-    GKO_ASSERT_MTX_NEAR(mtx, dmtx, 1e-14);
+    GKO_ASSERT_MTX_NEAR(mtx, dmtx, r<vtype>::value);
 }
 
 
@@ -358,8 +382,5 @@ TEST_F(Coo, OutplaceAbsoluteMatrixIsEquivalentToRef)
     auto abs_mtx = mtx->compute_absolute();
     auto dabs_mtx = dmtx->compute_absolute();
 
-    GKO_ASSERT_MTX_NEAR(abs_mtx, dabs_mtx, 1e-14);
+    GKO_ASSERT_MTX_NEAR(abs_mtx, dabs_mtx, r<vtype>::value);
 }
-
-
-}  // namespace
